requirements.txt
absl-py==2.0.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
astroid==2.15.6
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
Babel==2.15.0
backcall==0.2.0
bleach==6.1.0
blinker==1.9.0
branca==0.6.0
cachetools==5.3.2
cassandra-driver==3.28.0
catboost==1.2.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
daal==2024.4.0
daal4py==2024.4.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
Faker==20.1.0
fastapi==0.117.1
fastjsonschema==2.18.0
filelock==3.17.0
Fiona==1.9.4.post1
Flask==3.0.0
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.39.4
gast==0.5.4
geographiclib==2.0
geomet==0.2.1.post1
geopandas==0.13.2
geopy==2.4.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.60.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.10.0
haversine==2.8.0
hilbertcurve==2.0.5
idna==3.4
imageio==2.28.1
ipykernel==6.23.2
ipympl==0.9.4
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.1.3
isort==5.12.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
july==0.1.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab_widgets==3.0.11
kaggle==1.6.17
kaleido==0.2.1
keras==2.15.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.2
libclang==16.0.6
lightgbm==4.3.0
loguru==0.7.2
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.2.0
mpld3==0.5.9
mpmath==1.3.0
narwhals==2.14.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.1.4
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.3.6
plotly==5.15.0
plotly-express==0.4.1
pmongo==1.0.16
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==22.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.4.0
Pygments==2.15.1
pylint==2.17.5
pymongo==4.15.1
pyparsing==3.0.9
pyproj==3.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
referencing==0.30.0
requests==2.30.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-base==0.6.1
scikit-image==0.20.0
scikit-learn==1.3.2
scikit-learn-intelex==2024.4.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
shapely==2.0.1
six==1.16.0
sklearn==0.0.post10
sktime==0.24.1
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
starlette==0.48.0
statsmodels==0.14.0
streamlit==1.52.2
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.4.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.37.0
virtualenv==20.29.2
watchdog==6.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.11
win32_setctime==1.2.0
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
